+++ Marlin/src/gcode/lcd/M117.cpp
@@ -26,28 +26,13 @@
 /**
  * M117: Set LCD Status Message
  */
-#if ENABLED(DWIN_CREALITY_LCD)
 
-  #include "../../lcd/dwin/e3v2/dwin.h"
+void GcodeSuite::M117() {
 
-  void GcodeSuite::M117() {
+  if (parser.string_arg && parser.string_arg[0])
+    ui.set_status(parser.string_arg);
+  else
+    ui.reset_status();
 
-    if (parser.string_arg && parser.string_arg[0])
-      Host_Print_Text(parser.string_arg);
-    else
-      Host_Print_Text((char*)"");
+}
 
-  }
-
-#else
-
-  void GcodeSuite::M117() {
-
-    if (parser.string_arg && parser.string_arg[0])
-      ui.set_status(parser.string_arg);
-    else
-      ui.reset_status();
-
-  }
-
-#endif
