+++ Marlin/src/gcode/sd/M24_M25.cpp
@@ -113,7 +113,7 @@ void GcodeSuite::M25() {
 
     TERN_(DGUS_LCD_UI_MKS, MKS_pause_print_move());
 
-    IF_DISABLED(DWIN_CREALITY_LCD, ui.reset_status());
+    ui.reset_status();
 
     #if ENABLED(HOST_ACTION_COMMANDS)
       TERN_(HOST_PROMPT_SUPPORT, host_prompt_open(PROMPT_PAUSE_RESUME, PSTR("Pause SD"), PSTR("Resume")));
