+++ Marlin/src/lcd/dwin/dwin.cpp
@@ -34,7 +34,7 @@
 
 #include "../../inc/MarlinConfig.h"
 
-#include "dwin_lcd.h"
+#include "dwin.h"
 #include <string.h> // for memset
 
 //#define DEBUG_OUT 1
@@ -95,7 +95,7 @@ bool DWIN_Handshake(void) {
   #endif
   LCD_SERIAL.begin(LCD_BAUDRATE);
   const millis_t serial_connect_timeout = millis() + 1000UL;
-  while (!LCD_SERIAL.connected() && PENDING(millis(), serial_connect_timeout)) { /*nada*/ }
+  while (!LCD_SERIAL && PENDING(millis(), serial_connect_timeout)) { /*nada*/ }
 
   size_t i = 0;
   DWIN_Byte(i, 0x00);
@@ -127,7 +127,7 @@ bool DWIN_Handshake(void) {
 void DWIN_Backlight_SetLuminance(const uint8_t luminance) {
   size_t i = 0;
   DWIN_Byte(i, 0x30);
-  DWIN_Byte(i, _MAX(luminance, 0x1F));
+  DWIN_Byte(i, luminance);
   DWIN_Send(i);
 }
 
@@ -161,12 +161,14 @@ void DWIN_Frame_Clear(const uint16_t color) {
 }
 
 // Draw a point
+//  color: Pixel segment color
 //  width: point width   0x01-0x0F
 //  height: point height 0x01-0x0F
 //  x,y: upper left point
-void DWIN_Draw_Point(uint8_t width, uint8_t height, uint16_t x, uint16_t y) {
+void DWIN_Draw_Point(uint16_t color, uint8_t width, uint8_t height, uint16_t x, uint16_t y) {
   size_t i = 0;
   DWIN_Byte(i, 0x02);
+  DWIN_Word(i, color);
   DWIN_Byte(i, width);
   DWIN_Byte(i, height);
   DWIN_Word(i, x);
@@ -206,6 +208,20 @@ void DWIN_Draw_Rectangle(uint8_t mode, uint16_t color,
   DWIN_Word(i, yEnd);
   DWIN_Send(i);
 }
+
+//Color: color
+//x/y: Upper-left coordinate of the first pixel
+void DWIN_Draw_DegreeSymbol(uint16_t Color, uint16_t x, uint16_t y)	{
+  	DWIN_Draw_Point(Color, 1, 1, x + 1, y);		               	
+  	DWIN_Draw_Point(Color, 1, 1, x + 2, y);	
+  	DWIN_Draw_Point(Color, 1, 1, x, y + 1);
+		DWIN_Draw_Point(Color, 1, 1, x + 3, y + 1);
+  	DWIN_Draw_Point(Color, 1, 1, x, y + 2);
+		DWIN_Draw_Point(Color, 1, 1, x + 3, y + 2);
+    DWIN_Draw_Point(Color, 1, 1, x + 1, y + 3);		               	
+  	DWIN_Draw_Point(Color, 1, 1, x + 2, y + 3);	
+}
+
 
 // Move a screen area
 //  mode: 0, circle shift; 1, translation
