+++ Marlin/src/lcd/dwin/rotary_encoder.cpp
@@ -30,13 +30,13 @@
 
 #include "../../inc/MarlinConfigPre.h"
 
-#if ENABLED(CREALITY_DWIN_EXTUI)
+#if ENABLED(DWIN_CREALITY_LCD)
 
 #include "rotary_encoder.h"
-#include "../../buttons.h"
+#include "../buttons.h"
 
 #include "../../MarlinCore.h"
-#include "../../marlinui.h"
+#include "../marlinui.h"
 #include "../../HAL/shared/Delay.h"
 
 #if HAS_BUZZER
