--- Marlin/src/lcd/extui/creality_dwin/creality_extui.cpp
@@ -1,90 +1,0 @@
-/****************************************************************************
- *   Written By Marcio Teixeira 2018 - Aleph Objects, Inc.                  *
- *                                                                          *
- *   This program is free software: you can redistribute it and/or modify   *
- *   it under the terms of the GNU General Public License as published by   *
- *   the Free Software Foundation, either version 3 of the License, or      *
- *   (at your option) any later version.                                    *
- *                                                                          *
- *   This program is distributed in the hope that it will be useful,        *
- *   but WITHOUT ANY WARRANTY; without even the implied warranty of         *
- *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the          *
- *   GNU General Public License for more details.                           *
- *                                                                          *
- *   To view a copy of the GNU General Public License, go to the following  *
- *   location: <https://www.gnu.org/licenses/>.                             *
- ****************************************************************************/
-
-/**
- * creality_dwin.cpp
- *
- * DWIN implementation for Marlin by Jacob Myers, 2021
- */
-
-#include "../../inc/MarlinConfigPre.h"
-
-#if BOTH(CREALITY_DWIN_EXTUI, EXTENSIBLE_UI)
-
-#include "creality_dwin.h"
-#include "../ui_api.h"
-
-namespace ExtUI {
-  void onStartup() {}
-  void onIdle() {}
-  void onPrinterKilled(PGM_P const error, PGM_P const component) {}
-  void onMediaInserted() {}
-  void onMediaError() {}
-  void onMediaRemoved() {}
-  void onPlayTone(const uint16_t frequency, const uint16_t duration) {}
-  void onPrintTimerStarted() {}
-  void onPrintTimerPaused() {}
-  void onPrintTimerStopped() {}
-  void onFilamentRunout(const extruder_t extruder) {}
-  void onUserConfirmRequired(const char * const msg) {}
-  void onStatusChanged(const char * const msg) {}
-
-  void onHomingStart() {}
-  void onHomingComplete() {}
-  void onPrintFinished() {}
-
-  void onFactoryReset() {}
-
-  void onStoreSettings(char *buff) {
-  }
-
-  void onLoadSettings(const char *buff) {
-  }
-
-  void onPostprocessSettings() {
-    // Called after loading or resetting stored settings
-  }
-
-  void onConfigurationStoreWritten(bool success) {}
-
-  void onConfigurationStoreRead(bool success) {}
-
-  #if HAS_MESH
-    void onMeshLevelingStart() {}
-
-    void onMeshUpdate(const int8_t xpos, const int8_t ypos, const_float_t zval) {
-      // Called when any mesh points are updated
-    }
-    void onMeshUpdate(const int8_t xpos, const int8_t ypos, const probe_state_t state) {
-      // Called to indicate a special condition
-    }
-  #endif
-
-  #if ENABLED(POWER_LOSS_RECOVERY)
-    void onPowerLossResume() { CrealityDWIN.Popup_Handler(Resume); }
-  #endif
-
-  #if HAS_PID_HEATING
-    void onPidTuning(const result_t rst) {
-    }
-  #endif
-
-  void onSteppersDisabled() {}
-  void onSteppersEnabled()  {}
-}
-
-#endif // CREALITY_DWIN_EXTUI && EXTENSIBLE_UI
