+++ Marlin/src/lcd/marlinui.cpp
@@ -47,11 +47,7 @@ MarlinUI ui;
 #endif
 
 #if ENABLED(DWIN_CREALITY_LCD)
-  #include "dwin/e3v2/dwin.h"
-#endif
-
-#if ENABLED(CREALITY_DWIN_EXTUI)
-  #include "extui/creality_dwin/creality_dwin.h"
+  #include "dwin/creality_dwin.h"
 #endif
 
 #if ENABLED(LCD_PROGRESS_BAR) && !IS_TFTGLCD_PANEL
@@ -1459,8 +1455,7 @@ void MarlinUI::update() {
     #endif
 
     TERN_(EXTENSIBLE_UI, ExtUI::onStatusChanged(status_message));
-    TERN_(DWIN_CREALITY_LCD, DWIN_StatusChanged(status_message));
-    TERN_(CREALITY_DWIN_EXTUI, CrealityDWIN.Update_Status(status_message));
+    TERN_(DWIN_CREALITY_LCD, CrealityDWIN.Update_Status(status_message));
   }
 
   #if ENABLED(STATUS_MESSAGE_SCROLLING)
