+++ Marlin/src/pins/lpc1768/pins_BTT_SKR_V1_4.h
@@ -267,7 +267,7 @@
  *               EXP2                                              EXP1
  */
 
-#if ANY(DWIN_CREALITY_LCD, CREALITY_DWIN_EXTUI)
+#if ENABLED(DWIN_CREALITY_LCD)
 
   // RET6 DWIN ENCODER LCD
   #define BTN_ENC                          P1_20
